requirements.txt
fastapi>=0.115.11
uvicorn>=0.34.0
pydantic>=2.6.1
pydantic-settings>=2.1.0
sqlalchemy>=2.0.27
redis>=5.0.1
prometheus-client>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6 